Moved tests assets for user defined fuctions from sql/v2/test to sql/v2/runtime
